backend/requirements.txt
flask
flask-cors
fastapi
uvicorn
requests
pydantic
python-dotenv
sqlalchemy
oracledb
pymongo
passlib[argon2]
python-jose
python-multipart